requirements.txt
eth_abi==2.1.1
pycryptodome==3.14.1
pysha3==1.0.2
requests==2.26.0
sha3==0.2.1
web3==5.24.0
